making get_coords more general.
get_coords borrows a 2d array and creates a new 2d array with a tuple containing the original 2d array's value T, row, and column.

unfortunately, T has to implement copy and can be slow if T is a big structure, but we can improve on that some other time by making T a reference if we need to. But that needs lifetimes and that can get annoying.

also some lines have been formatted due to format on save.